fix: reduce inbox API calls
