+++ api/useInboxItem.ts
@@ -1,9 +1,8 @@
 import { type Schema } from "@/amplify/data/resource";
-import { SerializerOutput } from "@/helpers/ui-notes-writer";
 import { generateClient } from "aws-amplify/data";
 import useSWR from "swr";
 import { handleApiErrors } from "./globals";
-import { InboxStatus, mapInbox } from "./useInbox";
+import { HandleMutationFn, mapInbox } from "./useInbox";
 const client = generateClient<Schema>();
 
 const fetchInboxItem = (itemId?: string) => async () => {
@@ -20,85 +19,20 @@ const fetchInboxItem = (itemId?: string) => async () => {
 const useInboxItem = (itemId?: string) => {
   const {
     data: inboxItem,
-    error: errorInboxItem,
-    isLoading: loadingInboxItem,
+    error,
+    isLoading,
     mutate,
   } = useSWR(`/api/inbox/${itemId}`, fetchInboxItem(itemId));
 
-  const updateNote = async (
-    id: string,
-    { json: note, hasOpenTasks }: SerializerOutput
-  ) => {
-    const { data, errors } = await client.models.Inbox.update({
-      id,
-      note: null,
-      formatVersion: 2,
-      noteJson: JSON.stringify(note),
-      hasOpenTasks: hasOpenTasks ? "true" : "false",
-    });
-    if (errors) handleApiErrors(errors, "Error updating inbox item");
-    if (!data) return;
-    const updated = mapInbox(data);
-    mutate(updated);
-    return data.id;
-  };
-
-  const moveInboxItemToProject = async (projectId: string) => {
-    if (!inboxItem) return;
-
-    const { data: activity, errors: activityErrors } =
-      await client.models.Activity.create({
-        finishedOn: inboxItem.createdAt.toISOString(),
-        formatVersion: 2,
-        notes: null,
-        notesJson: JSON.stringify(inboxItem.note),
-        hasOpenTasks: inboxItem.hasOpenTasks ? "true" : "false",
-      });
-    if (activityErrors)
-      return handleApiErrors(
-        activityErrors,
-        "Error creating activity with inbox notes"
-      );
-    if (!activity) return;
-
-    const { errors: projectActivityErrors } =
-      await client.models.ProjectActivity.create({
-        activityId: activity.id,
-        projectsId: projectId,
-      });
-    if (projectActivityErrors)
-      return handleApiErrors(
-        projectActivityErrors,
-        "Error linking activity with project"
-      );
-
-    const { data, errors } = await client.models.Inbox.update({
-      id: inboxItem.id,
-      status: "done",
-      movedToActivityId: activity.id,
-    });
-
-    if (errors)
-      return handleApiErrors(errors, "Error updating status of inbox item");
-    return data?.id;
-  };
-
-  const updateStatus = async (id: string, status: InboxStatus) => {
-    if (inboxItem) mutate({ ...inboxItem, status }, false);
-    const { data, errors } = await client.models.Inbox.update({ id, status });
-    if (errors) handleApiErrors(errors, "Can't update status");
-    if (!data) return;
-    mutate(mapInbox(data));
-    return data.id;
+  const handleMutation: HandleMutationFn = (item, callApi = true) => {
+    mutate(item, callApi);
   };
 
   return {
     inboxItem,
-    errorInboxItem,
-    loadingInboxItem,
-    updateNote,
-    moveInboxItemToProject,
-    updateStatus,
+    error,
+    isLoading,
+    mutate: handleMutation,
   };
 };
