+++ components/inbox/WorkflowItem.tsx
@@ -1,4 +1,5 @@
-import useInboxItem from "@/api/useInboxItem";
+import { HandleMutationFn, Inbox } from "@/api/useInbox";
+import useInboxWorkflow from "@/api/useInboxWorkflow";
 import {
   getEditorContentAndTaskData,
   TWithGetJsonFn,
@@ -8,6 +9,7 @@ import { useRouter } from "next/router";
 import { FC, useEffect, useState } from "react";
 import NotesWriter from "../ui-elements/notes-writer/NotesWriter";
 import { Button } from "../ui/button";
+import { Skeleton } from "../ui/skeleton";
 import {
   getPreviousStatusByStatus,
   getWorkflowStepByStatus,
@@ -16,20 +18,27 @@ import {
 } from "./workflow";
 
 type WorkFlowItemProps = {
-  inboxItemId: string;
+  inboxItem?: Inbox;
+  isLoading?: boolean;
   forwardUrl?: string;
+  mutate: HandleMutationFn;
 };
 
-const WorkFlowItem: FC<WorkFlowItemProps> = ({ inboxItemId, forwardUrl }) => {
+const WorkFlowItem: FC<WorkFlowItemProps> = ({
+  inboxItem,
+  forwardUrl,
+  isLoading,
+  mutate,
+}) => {
   const router = useRouter();
-  const { inboxItem, updateNote, updateStatus, moveInboxItemToProject } =
-    useInboxItem(inboxItemId);
   const [step, setStep] = useState(
     getWorkflowStepByStatus(workflow, inboxItem?.status)
   );
   const [prevStatus, setPrevStatus] = useState(
     getPreviousStatusByStatus(workflow, inboxItem?.status || "new")
   );
+  const { updateNote, updateStatus, moveInboxItemToProject } =
+    useInboxWorkflow(mutate);
 
   useEffect(() => {
     if (!inboxItem) return;
@@ -38,16 +47,18 @@ const WorkFlowItem: FC<WorkFlowItemProps> = ({ inboxItemId, forwardUrl }) => {
   }, [inboxItem]);
 
   const handleUpdate = (editor: TWithGetJsonFn) => {
+    if (!inboxItem) return;
     debouncedOnChangeInboxNote(
-      inboxItemId,
+      inboxItem.id,
       getEditorContentAndTaskData(editor, () => {}),
-      updateNote
+      updateNote(inboxItem)
     );
   };
 
   const goBack = async () => {
     if (!prevStatus) return;
-    await updateStatus(inboxItemId, prevStatus);
+    if (!inboxItem) return;
+    await updateStatus(inboxItem, prevStatus);
   };
 
   const startProcessingItem = async (
@@ -57,10 +68,10 @@ const WorkFlowItem: FC<WorkFlowItemProps> = ({ inboxItemId, forwardUrl }) => {
     if (!inboxItem) return;
     if (!response.nextStep) return;
     if (projectId) {
-      const result = await moveInboxItemToProject(projectId);
+      const result = await moveInboxItemToProject(inboxItem, projectId);
       if (!result) return;
     } else {
-      const result = await updateStatus(inboxItemId, response.nextStep.status);
+      const result = await updateStatus(inboxItem, response.nextStep.status);
       if (!result) return;
     }
 
@@ -68,26 +79,28 @@ const WorkFlowItem: FC<WorkFlowItemProps> = ({ inboxItemId, forwardUrl }) => {
     if (response.nextStep.toHome) router.replace("/inbox");
   };
 
-  return !inboxItem ? (
-    "Loading inbox item…"
+  return isLoading ? (
+    <Skeleton className="w-full h-8" />
   ) : !step ? (
     "Loading workflow…"
   ) : (
-    <div>
-      {inboxItem.status !== "new" && (
-        <Button size="sm" onClick={goBack}>
-          Return to previous step
-        </Button>
-      )}
-      {step.component && (
-        <step.component
-          question={step.question}
-          responses={step.responses}
-          action={startProcessingItem}
-        />
-      )}
-      <NotesWriter notes={inboxItem?.note} saveNotes={handleUpdate} />
-    </div>
+    inboxItem && (
+      <div>
+        {inboxItem.status !== "new" && (
+          <Button size="sm" onClick={goBack}>
+            Return to previous step
+          </Button>
+        )}
+        {step.component && (
+          <step.component
+            question={step.question}
+            responses={step.responses}
+            action={startProcessingItem}
+          />
+        )}
+        <NotesWriter notes={inboxItem?.note} saveNotes={handleUpdate} />
+      </div>
+    )
   );
 };
