+++ docs/releases/next.md
@@ -6,14 +6,14 @@
 - Die App hat stetig den Inbox API Endpunkt aufgerufen, nur damit Anwender einen neuen Eintrag zu der Inbox hinzufügen können. Das passiert nun nicht mehr.
 - Die Einträge der Inbox wurden nicht korrekt dargestellt, weil das Feld `hasOpenTasks` für alte Einträge nicht gesetzt war.
 - Personeninformationen bereits im Accordion Untertitel anzeigen (Unternehmen und Aussprache des Namens).
+- Die Anzahl der API Aufrufe für Einträge in der Inbox reduziert.
 
 ## In Arbeit
 
-- Die Anzahl der API Aufrufe für Einträge in der Inbox reduziert.
+- In Meetings werden beim Laden nicht die nächsten Aufgaben korrekt dargestellt. Wenn man eine Notiz bearbeitet und dadurch eine Aufgabe hinzufügt, werden auch die nächsten Aufgaben korrekt dargestellt, nicht aber schon beim Laden.
 
 ## Geplant
 
-- In Meetings werden beim Laden nicht die nächsten Aufgaben korrekt dargestellt. Wenn man eine Notiz bearbeitet und dadurch eine Aufgabe hinzufügt, werden auch die nächsten Aufgaben korrekt dargestellt, nicht aber schon beim Laden.
 - In Projekten werden beim Laden nicht die nächsten Aufgaben korrekt dargestellt. Wenn man eine Notiz bearbeitet und dadurch eine Aufgabe hinzufügt, werden auch die nächsten Aufgaben korrekt dargestellt, nicht aber schon beim Laden.
 
 ## Nice to have
